Fixed java surrounding pair
