+++ src/languages/getTextFragmentExtractor.ts
@@ -1,11 +1,13 @@
 import { SyntaxNode } from "web-tree-sitter";
 import { SelectionWithEditor } from "../typings/Types";
-import { textFragmentExtractor as jsonTextFragmentExtractor } from "./json";
+import { stringTextFragmentExtractor as jsonStringTextFragmentExtractor } from "./json";
+import { stringTextFragmentExtractor as javaStringTextFragmentExtractor } from "./java";
 import { UnsupportedLanguageError } from "../errors";
 import { Range } from "vscode";
 import { SupportedLanguageId } from "./constants";
 import { getNodeRange, makeRangeFromPositions } from "../util/nodeSelectors";
-import { getNodeMatcher } from "./index";
+import { getNodeMatcher } from "./getNodeMatcher";
+import { notSupported } from "../util/nodeMatchers";
 
 export type TextFragmentExtractor = (
   node: SyntaxNode,
@@ -13,25 +15,45 @@ export type TextFragmentExtractor = (
 ) => Range | null;
 
 function constructDefaultTextFragmentExtractor(
+  languageId: SupportedLanguageId,
+  stringTextFragmentExtractor?: TextFragmentExtractor
+): TextFragmentExtractor {
+  const commentNodeMatcher = getNodeMatcher(languageId, "comment", false);
+  stringTextFragmentExtractor =
+    stringTextFragmentExtractor ??
+    constructDefaultStringTextFragmentExtractor(languageId);
+
+  return (node: SyntaxNode, selection: SelectionWithEditor) => {
+    const stringTextFragment = stringTextFragmentExtractor!(node, selection);
+
+    if (stringTextFragment != null) {
+      return stringTextFragment;
+    }
+
+    if (
+      commentNodeMatcher !== notSupported &&
+      commentNodeMatcher(selection, node) != null
+    ) {
+      return getNodeRange(node);
+    }
+
+    return null;
+  };
+}
+
+function constructDefaultStringTextFragmentExtractor(
   languageId: SupportedLanguageId
 ): TextFragmentExtractor {
   const stringNodeMatcher = getNodeMatcher(languageId, "string", false);
-  const commentNodeMatcher = getNodeMatcher(languageId, "comment", false);
 
   return (node: SyntaxNode, selection: SelectionWithEditor) => {
-    const isStringNode = stringNodeMatcher(selection, node) != null;
+    if (stringNodeMatcher(selection, node) != null) {
+      const children = node.children;
 
-    if (isStringNode || commentNodeMatcher(selection, node) != null) {
-      if (isStringNode) {
-        const children = node.children;
-
-        return makeRangeFromPositions(
-          children[0].endPosition,
-          children[children.length - 1].startPosition
-        );
-      } else {
-        return getNodeRange(node);
-      }
+      return makeRangeFromPositions(
+        children[0].endPosition,
+        children[children.length - 1].startPosition
+      );
     }
 
     return null;
@@ -57,11 +79,20 @@ const textFragmentExtractors: Record<
   c: constructDefaultTextFragmentExtractor("c"),
   cpp: constructDefaultTextFragmentExtractor("cpp"),
   csharp: constructDefaultTextFragmentExtractor("csharp"),
-  java: constructDefaultTextFragmentExtractor("java"),
+  java: constructDefaultTextFragmentExtractor(
+    "java",
+    javaStringTextFragmentExtractor
+  ),
   javascript: constructDefaultTextFragmentExtractor("javascript"),
   javascriptreact: constructDefaultTextFragmentExtractor("javascriptreact"),
-  jsonc: jsonTextFragmentExtractor,
-  json: jsonTextFragmentExtractor,
+  jsonc: constructDefaultTextFragmentExtractor(
+    "jsonc",
+    jsonStringTextFragmentExtractor
+  ),
+  json: constructDefaultTextFragmentExtractor(
+    "json",
+    jsonStringTextFragmentExtractor
+  ),
   python: constructDefaultTextFragmentExtractor("python"),
   typescript: constructDefaultTextFragmentExtractor("typescript"),
   typescriptreact: constructDefaultTextFragmentExtractor("typescriptreact"),
