+++ src/languages/index.ts
@@ -1,105 +1,7 @@
-import { SyntaxNode } from "web-tree-sitter";
-import { notSupported } from "../util/nodeMatchers";
-import { selectionWithEditorFromRange } from "../util/selectionUtils";
-import {
-  NodeMatcher,
-  NodeMatcherValue,
-  ScopeType,
-  SelectionWithEditor,
-} from "../typings/Types";
-import cpp from "./cpp";
-import csharp from "./csharp";
-import java from "./java";
-import { patternMatchers as json } from "./json";
-import python from "./python";
-import typescript from "./typescript";
-import { UnsupportedLanguageError } from "../errors";
 import { SupportedLanguageId, supportedLanguageIds } from "./constants";
-
-const languageMatchers: Record<
-  SupportedLanguageId,
-  Record<ScopeType, NodeMatcher>
-> = {
-  c: cpp,
-  cpp: cpp,
-  csharp: csharp,
-  java,
-  javascript: typescript,
-  javascriptreact: typescript,
-  json,
-  jsonc: json,
-  python,
-  typescript,
-  typescriptreact: typescript,
-};
 
 export function isLanguageSupported(
   languageId: string
 ): languageId is SupportedLanguageId {
   return languageId in supportedLanguageIds;
-}
-
-export function getNodeMatcher(
-  languageId: string,
-  scopeType: ScopeType,
-  includeSiblings: boolean
-): NodeMatcher {
-  const matchers = languageMatchers[languageId as SupportedLanguageId];
-
-  if (matchers == null) {
-    throw new UnsupportedLanguageError(languageId);
-  }
-
-  const matcher = matchers[scopeType];
-
-  if (matcher == null) {
-    return notSupported;
-  }
-
-  if (includeSiblings) {
-    return matcherIncludeSiblings(matcher);
-  }
-
-  return matcher;
-}
-
-function matcherIncludeSiblings(matcher: NodeMatcher): NodeMatcher {
-  return (
-    selection: SelectionWithEditor,
-    node: SyntaxNode
-  ): NodeMatcherValue[] | null => {
-    let matches = matcher(selection, node);
-    if (matches == null) {
-      return null;
-    }
-    matches = matches.flatMap((match) =>
-      iterateNearestIterableAncestor(
-        match.node,
-        selectionWithEditorFromRange(selection, match.selection.selection),
-        matcher
-      )
-    ) as NodeMatcherValue[];
-    if (matches.length > 0) {
-      return matches;
-    }
-    return null;
-  };
-}
-
-function iterateNearestIterableAncestor(
-  node: SyntaxNode,
-  selection: SelectionWithEditor,
-  nodeMatcher: NodeMatcher
-) {
-  let parent: SyntaxNode | null = node.parent;
-  while (parent != null) {
-    const matches = parent!.namedChildren
-      .flatMap((sibling) => nodeMatcher(selection, sibling))
-      .filter((match) => match != null) as NodeMatcherValue[];
-    if (matches.length > 0) {
-      return matches;
-    }
-    parent = parent.parent;
-  }
-  return [];
 }
