+++ src/languages/java.ts
@@ -5,7 +5,13 @@ import {
   conditionMatcher,
   trailingMatcher,
 } from "../util/nodeMatchers";
-import { NodeMatcherAlternative, ScopeType } from "../typings/Types";
+import {
+  NodeMatcherAlternative,
+  ScopeType,
+  SelectionWithEditor,
+} from "../typings/Types";
+import { getNodeRange } from "../util/nodeSelectors";
+import { SyntaxNode } from "web-tree-sitter";
 
 // Generated by the following command:
 // > curl https://raw.githubusercontent.com/tree-sitter/tree-sitter-java/master/src/node-types.json | jq '[.[] | select(.type == "statement" or .type == "declaration") | .subtypes[].type]'
@@ -70,4 +76,28 @@ const nodeMatchers: Partial<Record<ScopeType, NodeMatcherAlternative>> = {
   argumentOrParameter: argumentMatcher("formal_parameters", "argument_list"),
 };
 
-export default createPatternMatchers(nodeMatchers);
+export const patternMatchers = createPatternMatchers(nodeMatchers);
+
+/**
+ * Extracts string text fragments in java.
+ *
+ * This is a hack to deal with the fact that java doesn't have
+ * quotation mark tokens as children of the string. Rather than letting
+ * the parse tree handle the quotation marks in java, we instead just
+ * let the textual surround handle them by letting it see the quotation
+ * marks. In other languages we prefer to let the parser handle the
+ * quotation marks in case they are more than one character long.
+ * @param node The node which might be a string node
+ * @param selection The selection from which to expand
+ * @returns The range of the string text or null if the node is not a string
+ */
+export function stringTextFragmentExtractor(
+  node: SyntaxNode,
+  selection: SelectionWithEditor
+) {
+  if (node.type === "string_literal") {
+    return getNodeRange(node);
+  }
+
+  return null;
+}
